requirements.txt
boto==2.43.0
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.4
django-allauth==0.29.0
django-bootstrap3==7.1.0
django-crispy-forms==1.6.1
django-sortedm2m==1.3.2
django-tinymce==2.4.0
Django-tinymce-filebrowser==0.2.1
django-toolbelt==0.0.1
gunicorn==19.6.0
oauthlib==2.0.1
Pillow==3.4.2
psycopg2==2.6.2
python-openid==2.2.5
python3-openid==3.0.10
requests==2.12.3
requests-oauthlib==0.7.0
sorl-thumbnail==12.3
static3==0.7.0
wheel==0.24.0
